Rename file news-item-post.vue to NewsItemPost.vue
